Conflicts:
	sagebrew/sagebrew/templates/base.html

+++ sagebrew/sb_registration/utils.py
@@ -1,9 +1,7 @@
 import os
 import shortuuid
 import hashlib
-import json
 import boto.ses
-import logging
 from boto.ses.exceptions import SESMaxSendingRateExceededError
 from datetime import date
 from django.conf import settings
@@ -16,8 +14,8 @@ from neomodel import DoesNotExist, CypherException
 from api.utils import spawn_task
 from plebs.tasks import create_pleb_task
 from plebs.neo_models import Pleb
+from sb_base.decorators import apply_defense
 
-logger = logging.getLogger('loggly_logs')
 
 
 def calc_age(birthday):
@@ -34,7 +32,7 @@ def calc_age(birthday):
 
 
 def create_address_long_hash(address):
-    if ("address2" in address):
+    if "street_additional" in address:
         address_string = "%s%s%s%s%s%s%f%f%s" % (address["primary_address"],
                                                  address["street_additional"],
                                                  address["city"],
@@ -180,7 +178,7 @@ def validate_school(school_name):
     pass
 
 
-def upload_image(folder_name, file_uuid):
+def upload_image(folder_name, file_uuid, file_location=None):
     '''
     Creates a connection to the s3 service then uploads the file which was
     passed
@@ -190,17 +188,26 @@ def upload_image(folder_name, file_uuid):
     :param file_uuid:
     :return:
     '''
-    file_path = '%s%s.%s' % (settings.TEMP_FILES, file_uuid, 'jpeg')
+    if file_location is None:
+        file_path = '%s%s.%s' % (settings.TEMP_FILES, file_uuid, 'jpg')
+    else:
+        file_path = '%s%s.%s' % (file_location, file_uuid, 'jpg')
 
     bucket = settings.AWS_STORAGE_BUCKET_NAME
     conn = connect_s3(settings.AWS_ACCESS_KEY_ID,
                       settings.AWS_SECRET_ACCESS_KEY)
     k = Key(conn.get_bucket(bucket))
-    key_string = "%s/%s.%s" % (folder_name, file_uuid, "jpeg")
+    key_string = "%s/%s.%s" % (folder_name, file_uuid, "jpg")
     k.key = key_string
     k.set_contents_from_filename(file_path)
     image_uri = k.generate_url(expires_in=259200)
-    os.remove(file_path)
+    # TODO This should be reviewed and updated. It is a quick fix for
+    # not deleting the test file in sb_posts/tests/images when running
+    # through tests. We should also be looking into just storing the image
+    # in memory or directly in s3 rather than bringing it locally onto the
+    # system.
+    if os.environ.get("CIRCLECI", "false") == "false":
+        os.remove(file_path)
     return image_uri
 
 
@@ -291,9 +298,7 @@ def verify_completed_registration(user):
     try:
         pleb = Pleb.nodes.get(email=user.email)
         return pleb.completed_profile_info
-    except (Pleb.DoesNotExist,DoesNotExist):
-        return False
-    except CypherException:
+    except (Pleb.DoesNotExist,DoesNotExist,CypherException):
         return False
 
 
@@ -310,12 +315,10 @@ def verify_verified_email(user):
         return pleb.email_verified
     except (Pleb.DoesNotExist, DoesNotExist):
         return False
-    except CypherException:
-        logger.critical({"exception": "cypher exception",
-                         "function": "verify_verified_email"})
-        return False
+    except CypherException as e:
+        return e
 
-
+@apply_defense
 def sb_send_email(to_email, subject, html_content):
     '''
     This function is used to send mail through the amazon ses service,
@@ -339,33 +342,20 @@ def sb_send_email(to_email, subject, html_content):
         return True
     except SESMaxSendingRateExceededError as e:
         return e
-    except Exception as e:
-        logger.exception(json.dumps({"function": sb_send_email.__name__,
-                                     "exception": "Unhandled Exception"}))
-        return e
 
-
+@apply_defense
 def create_user_util(first_name, last_name, email, password, username=None):
-    try:
-        if username is None:
-            username = shortuuid.uuid()
-        user = User.objects.create_user(first_name=first_name,
-                                        last_name=last_name,
-                                        email=email,
-                                        password=password,
-                                        username=username)
-        user.save()
-        res = spawn_task(task_func=create_pleb_task,
-                         task_param={"user_instance": user})
-        if res is not None:
-            return {"task_id": res, "username": user.username}
-        else:
-            logger.critical(json.dumps({"function": create_user_util.__name__,
-                          "exception": "res is None"}))
-            return False
-    except Exception as e:
-        logger.exception(json.dumps({"function": create_user_util.__name__,
-                                     "exception": "Unhandled Exception"}))
-        # TODO should this return the exception? Think most things rely on it
-        # being False if something goes wrong.
-        return False
+    if username is None:
+        username = str(shortuuid.uuid())
+    user = User.objects.create_user(first_name=first_name,
+                                    last_name=last_name,
+                                    email=email,
+                                    password=password,
+                                    username=username)
+    user.save()
+    res = spawn_task(task_func=create_pleb_task,
+                     task_param={"user_instance": user})
+    if isinstance(res, Exception) is True:
+        return res
+    else:
+        return {"task_id": res, "username": user.username}
